+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- add `core/first-identity`, `core/second-identity`
+
 ## [1.0.10] — 2022-02-14
 
 ## [1.0.9] — 2022-02-14